+++ app/routes/authenticated/thread.ts
@@ -1,10 +1,12 @@
 import { service } from '@ember/service';
-import RSVP, { reject } from 'rsvp';
+import { reject } from 'rsvp';
 import ThreadController from 'potber-client/controllers/authenticated/thread';
 import Thread from 'potber-client/models/thread';
 import CustomStore from 'potber-client/services/custom-store';
 import NewsfeedService from 'potber-client/services/newsfeed';
 import SlowRoute from '../slow';
+import ApiService from 'potber-client/services/api';
+import { Bookmark } from 'potber-client/services/api/models/bookmark';
 
 interface Params {
   TID: string;
@@ -17,9 +19,11 @@ interface Params {
 export interface ThreadRouteModel {
   thread: Thread;
   lastReadPost: string;
+  bookmark?: Bookmark;
 }
 
 export default class ThreadRoute extends SlowRoute {
+  @service declare api: ApiService;
   @service declare store: CustomStore;
   @service declare newsfeed: NewsfeedService;
 
@@ -65,10 +69,16 @@ export default class ThreadRoute extends SlowRoute {
           },
         },
       });
-      return RSVP.hash({
+      // We will also check whether the thread has been bookmarked. If it has, also load that bookmark.
+      const bookmarks = await this.api.findAllBookmarks();
+      const bookmark = bookmarks.find(
+        (bookmark) => bookmark.thread.id === thread.id,
+      );
+      return {
         thread,
         lastReadPost: lastReadPost,
-      } as ThreadRouteModel);
+        bookmark,
+      } as ThreadRouteModel;
     } catch (error: any) {
       if (error.message === 'not-found') {
         return null;